feat: separate core logic from bridge logic
