+++ libs/coin-modules/coin-xrp/src/bridge/broadcast.ts
@@ -1,0 +1,11 @@
+import { AccountBridge } from "@ledgerhq/types-live";
+import { patchOperationWithHash } from "@ledgerhq/coin-framework/operation";
+import { broadcast as broadcastLogic } from "../logic";
+import { Transaction } from "../types";
+
+export const broadcast: AccountBridge<Transaction>["broadcast"] = async ({
+  signedOperation: { signature, operation },
+}) => {
+  const hash = await broadcastLogic(signature);
+  return patchOperationWithHash(operation, hash);
+};
