+++ libs/coin-modules/coin-xrp/src/logic/index.ts
@@ -1,0 +1,6 @@
+export * from "./logic";
+
+import craftTransaction from "./craftTransaction";
+import broadcast from "./broadcast";
+
+export { broadcast, craftTransaction };
